Cleanup and optimizations in geometry
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4643 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/PhysicsLX56.cpp
@@ -717,7 +717,7 @@ public:
 				continue;
 			if(y>=mh) {
 				if( cClient->getGameLobby()->features[FT_InfiniteMap] )
-					bonus->pos().y =- mh;
+					bonus->pos().y =- (float)mh;
 				else
 					colideBonus(bonus, x,y);
 				return;
